+++ assembler/src/generator/expression.rs
@@ -22,22 +22,18 @@ impl Spanned<&Expression> {
                     .wrapping_mul((**b).as_ref().eval(symbol_table)?));
             }
             Expression::Div(a, b) => {
-                let denominator = (**b).as_ref().eval(symbol_table)? as i32;
+                let denominator = (**b).as_ref().eval(symbol_table)?;
                 if denominator == 0 {
                     return Err(Error::new("divison by 0 is undefined", b.span));
                 }
-                return Ok(
-                    ((**a).as_ref().eval(symbol_table)? as i32).wrapping_div(denominator) as u32,
-                );
+                return Ok((**a).as_ref().eval(symbol_table)?.wrapping_div(denominator));
             }
             Expression::Remainder(a, b) => {
-                let denominator = (**b).as_ref().eval(symbol_table)? as i32;
+                let denominator = (**b).as_ref().eval(symbol_table)?;
                 if denominator == 0 {
                     return Err(Error::new("divison by 0 is undefined", b.span));
                 }
-                return Ok(
-                    ((**a).as_ref().eval(symbol_table)? as i32).wrapping_rem(denominator) as u32,
-                );
+                return Ok((**a).as_ref().eval(symbol_table)?.wrapping_rem(denominator));
             }
             Expression::Add(a, b) => {
                 return Ok((**a)
